Moved tests pertaining to User Interaction out of GameTest and into CommandLineUITest
