+++ GUI/src/javafxgui/FakeController.java
@@ -1,2 +1,4 @@
+package javafxgui;
+
 public class FakeController  {
 }
